requirements.txt
torch
torchvision
numpy
pandas
pillow
scikit-learn
matplotlib
tqdm
utm
pillow-heif